requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
basicsr==1.4.2
bitarray==2.8.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
dataclasses-json==0.6.1
decorator==4.4.2
diffusers==0.21.4
dlib==19.24.2
edge-tts==6.1.8
exceptiongroup==1.1.3
face-alignment==1.3.5
facexlib==0.3.0
fairseq==0.12.2
faiss-cpu==1.7.4
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
filterpy==1.4.5
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gfpgan==1.3.8
google-auth==2.23.2
google-auth-oauthlib==1.0.0
gradio==3.47.1
gradio_client==0.6.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
hydra-core==1.0.7
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kornia==0.7.0
langchain==0.0.311
langsmith==0.0.43
lazy_loader==0.3
librosa==0.10.1
lit==17.0.2
llvmlite==0.41.0
lmdb==1.4.1
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.0.6
openai==0.28.1
openai-whisper==20230918
opencv-python==4.8.1.78
orjson==3.9.7
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pooch==1.7.0
portalocker==2.8.2
praat-parselmouth==0.4.3
proglog==0.1.10
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
pyworld==0.3.4
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.9
sacrebleu==2.3.1
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tb-nightly==2.14.0a20230808
tenacity==8.2.3
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.3.3
tokenizers==0.14.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.34.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==3.0.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
